合并erp表 接口数据筛选功能

+++ src/main/java/com/benewake/system/entity/vo/QueryViewParams.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 @Data
 public class QueryViewParams {
-    Long tableId;
+    Integer tableId;
     Long viewId;
     List<ViewColParam> cols;
 }
